Refactor: 알림 비동기 처리
