+++ src/test/java/com/dangdangsalon/domain/notification/service/NotificationSchedulerTest.java
@@ -6,6 +6,7 @@ import com.dangdangsalon.domain.estimate.request.entity.EstimateRequest;
 import com.dangdangsalon.domain.groomerprofile.entity.GroomerProfile;
 import com.dangdangsalon.domain.notification.dto.NotificationDto;
 import com.dangdangsalon.domain.notification.dto.ReviewNotificationDto;
+import com.dangdangsalon.domain.notification.scheduler.NotificationScheduler;
 import com.dangdangsalon.domain.user.entity.User;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
